add Enemy class

+++ src/PlayerTest.hpp
@@ -29,14 +29,14 @@ namespace godot
         void _ready() override;
         void _physics_process(double delta) override;
 
-        void set_speed(double new_speed) { speed = new_speed; }
-        double get_speed() { return speed; }
+        void set_speed(int new_speed) { speed = new_speed; }
+        int get_speed() { return speed; }
 
         void set_jump_force(int new_jump_force) { jump_force = new_jump_force; }
         int get_jump_force() { return jump_force; }
 
-        void set_gravity(double new_gravity) { gravity = new_gravity; }
-        double get_gravity() { return gravity; }
+        void set_gravity(int new_gravity) { gravity = new_gravity; }
+        int get_gravity() { return gravity; }
     };
 }
